backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-mail==0.9.1
python-dotenv==1.0.1
requests==2.31.0
openai==1.12.0
boto3==1.37.16
redis==5.2.1
dramatiq==1.17.1
prometheus-client==0.21.1
Werkzeug==3.0.1
python-dateutil==2.8.2
urllib3==2.3.0 